% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GU-zhangjt.R
\name{calcKerMat_Cen}
\alias{calcKerMat_Cen}
\title{Similarity Matrix Calculation.}
\usage{
calcKerMat_Cen(X)
}
\arguments{
\item{X}{the data matrix, each row represents an observation.}
}
\value{
the centered similarity matrix.
}
\description{
\code{calcKerMat_Cen} calculates a centered similarity matrix for a data
matrix.
}
\examples{
library(MASS)
n = 50
p = 100
k = 15
sigmax = diag(rep(0.5,k)) + matrix(0.5,k,k)
x = mvrnorm(n, rep(0,k), sigmax)
Kx = calcKerMat_Cen(x)
}
