% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGU.R
\name{CGU9}
\alias{CGU9}
\title{Significance calculation for the combined generalized U statistc.}
\usage{
CGU9(mat0)
}
\arguments{
\item{mat0}{a matrix of p values to be combined.}
}
\value{
the p value of the combined generalized U statistic.
}
\description{
Significance calculation for the combined generalized U statistc.
}
\examples{
library(MASS)
n = 50
p = 100
k = 15
sigmax = diag(rep(0.5,k)) + matrix(0.5,k,k)
sigmay = diag(rep(1,p))
for(i in 1:p){
  for(j in 1:p){
    sigmay[i,j] = 0.5^abs(i-j)
  }
}
r1 = 0.05
beta0 = r1*matrix(rbinom(k*p,1,0.9), k, p)
x = mvrnorm(n, rep(0,k), sigmax)
y = x\%*\%beta0 + mvrnorm(n, rep(0,p), sigmay)
Kx = calcKerMat_Cen(x)
Ky = calcKerMat_Cen(y)
seq1 = c(0.5,1,2)
mat1 = zhangjtMat(Kx,Ky,seq1)
CGU9(mat1)
}
