% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GU-zhangjt.R
\name{zhangjtAppro}
\alias{zhangjtAppro}
\title{Significance Calculation for the Chi-squared-type Mixtures.}
\usage{
zhangjtAppro(E, G)
}
\arguments{
\item{E}{the similarity matrix of one of the variates or multivariates.}

\item{G}{the similarity matrix of the other one of the variates or
multivariates.}
}
\value{
the p value of the generalized U statistic, i.e.,
 $\frac 1n \text {tr}(EG)$, where $n$ is the sample size.
}
\description{
\code{zhangjtAppro} calculates the p value of the chi-squared-type mixtures
via the method of Jinting Zhang (2013).
}
\examples{
library(MASS)
n = 50
p = 100
k = 15
sigmax = diag(rep(0.5,k)) + matrix(0.5,k,k)
sigmay = diag(rep(1,p))
for(i in 1:p){
  for(j in 1:p){
    sigmay[i,j] = 0.5^abs(i-j)
  }
}
r1 = 0.05
beta0 = r1*matrix(rbinom(k*p,1,0.9), k, p)
x = mvrnorm(n, rep(0,k), sigmax)
y = x\%*\%beta0 + mvrnorm(n, rep(0,p), sigmay)
Kx = calcKerMat_Cen(x)
Ky = calcKerMat_Cen(y)
zhangjtAppro(Kx,Ky)
}
