% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGU.R
\name{zhangjtMat}
\alias{zhangjtMat}
\title{Significance Calculation of A Set of Generalized U Statistics.}
\usage{
zhangjtMat(E, G, seq0)
}
\arguments{
\item{E}{the similarity matrix of one of the variates or multivariates.}

\item{G}{the similarity matrix of the other one of the variates or
multivariates.}

\item{seq0}{a set of powers for the similarity matrixes.}
}
\value{
a matrix of p values.
}
\description{
\code{zhangjtMat} calculates a matrix of p values for a set of generatlized U
statistics, with each element corresponding to a generalized U statistic.
}
\examples{
library(MASS)
n = 50
p = 100
k = 15
sigmax = diag(rep(0.5,k)) + matrix(0.5,k,k)
sigmay = diag(rep(1,p))
for(i in 1:p){
  for(j in 1:p){
    sigmay[i,j] = 0.5^abs(i-j)
  }
}
r1 = 0.05
beta0 = r1*matrix(rbinom(k*p,1,0.9), k, p)
x = mvrnorm(n, rep(0,k), sigmax)
y = x\%*\%beta0 + mvrnorm(n, rep(0,p), sigmay)
Kx = calcKerMat_Cen(x)
Ky = calcKerMat_Cen(y)
seq1 = c(0.5,1,2)
zhangjtMat(Kx,Ky,seq1)
}
